refactor(backend): create services folder in posts module
