+++ packages/backend/src/modules/posts/services/index.ts
@@ -1,0 +1,2 @@
+export * from './GetPostsQuery';
+export * from './PostsService';
